Subject: Initial Commit
Description:
The goal of this project is to showcase some of the coding ideas and styles I like to utilize.  I'm a fan of Robert Martin's ideas on clean architecure which I hope is apperant. In this project I'm taking some of the code I've written for a class for colleaues and turing into a more extensible/maintainable model.

References:
[1] https://blog.cleancoder.com/uncle-bob/2012/08/13/the-clean-architecture.html
[2] https://www.amazon.com/Clean-Architecture-Craftsmans-Software-Structure/dp/0134494164/ref=asc_df_0134494164/?tag=hyprod-20&linkCode=df0&hvadid=312091457223&hvpos=1o1&hvnetw=g&hvrand=4861126156852357354&hvpone=&hvptwo=&hvqmt=&hvdev=c&hvdvcmdl=&hvlocint=&hvlocphy=9030030&hvtargid=pla-423658477418&psc=1&tag=&ref=&adgrpid=62820903995&hvpone=&hvptwo=&hvadid=312091457223&hvpos=1o1&hvnetw=g&hvrand=4861126156852357354&hvqmt=&hvdev=c&hvdvcmdl=&hvlocint=&hvlocphy=9030030&hvtargid=pla-423658477418